Abstract repeated/ rudimentary code into functions
